[bert_squeeze] - refactor: update distiller classes for sequence classification task
 - Rename distiller.py to sequence_classification_distiller.py
 - Refactor BaseDistiller to BaseSequenceClassificationDistiller
 - Modify Distiller class to SequenceClassificationDistiller
 - Create SequenceClassificationParallelDistiller class for multilingual distillation
 - Update assistant configs with adjusted distiller class names
