+++ bert_squeeze/distillation/sequence_classification_distiller.py
@@ -1,5 +1,5 @@
 import logging
-from typing import Any, Dict, List, Tuple, Union
+from typing import Dict, Tuple, Union, Any, List
 
 import lightning.pytorch as pl
 import matplotlib.pyplot as plt
@@ -7,23 +7,21 @@ import numpy as np
 import seaborn as sns
 import torch
 import torch.nn.functional as F
-from omegaconf import DictConfig, ListConfig
+from omegaconf import DictConfig
 from overrides import overrides
 from torch.nn import CrossEntropyLoss
-from torch.optim.lr_scheduler import ReduceLROnPlateau
-from transformers import AdamW
 from transformers.modeling_outputs import SequenceClassifierOutput
 
-from ..utils.losses import LabelSmoothingLoss
-from ..utils.losses.distillation_losses import KLDivLoss
-from ..utils.optimizers import BertAdam
-from ..utils.scorers import BaseSequenceClassificationScorer
-from ..utils.types import DistillationLoss
+from bert_squeeze.distillation.base_distiller import BaseDistiller
+from bert_squeeze.utils.losses import LabelSmoothingLoss
+from bert_squeeze.utils.losses.distillation_losses import KLDivLoss
+from bert_squeeze.utils.scorers import BaseSequenceClassificationScorer
+from bert_squeeze.utils.types import DistillationLoss
 
 
-class BaseDistiller(pl.LightningModule):
+class BaseSequenceClassificationDistiller(BaseDistiller):
     """
-    Base Lightning module to extend to perform distillation.
+    Lightning module to distil a given teacher model into a given student one for sequence classification task.
 
     Args:
         teacher (Union["pl.LightningModule", "torch.nn.Module"]):
@@ -37,25 +35,14 @@ class BaseDistiller(pl.LightningModule):
     """
 
     def __init__(
-        self,
-        teacher: Union["pl.LightningModule", "torch.nn.Module"],
-        student: Union[pl.LightningModule, torch.nn.Module],
-        training_config: DictConfig,
-        teacher_checkpoint: str = None,
-        **kwargs,
+            self,
+            teacher: Union["pl.LightningModule", "torch.nn.Module"],
+            student: Union["pl.LightningModule", "torch.nn.Module"],
+            training_config: DictConfig,
+            teacher_checkpoint: str = None,
+            **kwargs,
     ):
-        super().__init__()
-        self.params = training_config
-        self.teacher = teacher
-        self.student = student
-        self.teacher_checkpoint = teacher_checkpoint
-
-        self.training_step_outputs = []
-        self.test_step_outputs = []
-        self.validation_step_outputs = []
-
-        self._set_objectives()
-        self._set_scorers()
+        super().__init__(teacher, student, training_config, teacher_checkpoint, **kwargs)
 
     def _set_objectives(self) -> None:
         """
@@ -78,7 +65,7 @@ class BaseDistiller(pl.LightningModule):
                 "is set to 0.0."
             )
         elif objective == "weighted" and all(
-            [w == 1.0 for w in self.params.get("class_weights", None)]
+                [w == 1.0 for w in self.params.get("class_weights", None)]
         ):
             logging.warning(
                 "You are using a weighted CrossEntropy but the class"
@@ -108,179 +95,10 @@ class BaseDistiller(pl.LightningModule):
         self.s_valid_scorer = BaseSequenceClassificationScorer(self.params.num_labels)
         self.s_test_scorer = BaseSequenceClassificationScorer(self.params.num_labels)
 
-    def _get_student_parameters(self) -> List[Dict]:
-        """
-        Method that defines the student's parameters to optimize.
-
-        Returns:
-            List[Dict]: group of parameters to optimize
-        """
-        no_decay = ['bias', 'gamma', 'beta', 'LayerNorm.weight']
-
-        if self.params.discriminative_learning:
-            if (
-                isinstance(self.params.learning_rates, ListConfig)
-                and len(self.params.learning_rates) > 1
-            ):
-                groups = [
-                    (f'layer.{i}.', self.params.learning_rates[i]) for i in range(12)
-                ]
-            else:
-                lr = (
-                    self.params.learning_rates[0]
-                    if isinstance(self.params.learning_rates, ListConfig)
-                    else self.params.learning_rates
-                )
-                groups = [
-                    (f'layer.{i}.', lr * pow(self.params.layer_lr_decay, 11 - i))
-                    for i in range(12)
-                ]
-
-            group_all = [f'layer.{i}.' for i in range(12)]
-            no_decay_optimizer_parameters, decay_optimizer_parameters = [], []
-            for g, l in groups:
-                no_decay_optimizer_parameters.append(
-                    {
-                        'params': [
-                            p
-                            for n, p in self.student.named_parameters()
-                            if not any(nd in n for nd in no_decay)
-                            and any(nd in n for nd in [g])
-                        ],
-                        'weight_decay_rate': self.params.weight_decay,
-                        'lr': l,
-                    }
-                )
-                decay_optimizer_parameters.append(
-                    {
-                        'params': [
-                            p
-                            for n, p in self.student.named_parameters()
-                            if any(nd in n for nd in no_decay)
-                            and any(nd in n for nd in [g])
-                        ],
-                        'weight_decay_rate': 0.0,
-                        'lr': l,
-                    }
-                )
-
-            group_all_parameters = [
-                {
-                    'params': [
-                        p
-                        for n, p in self.student.named_parameters()
-                        if not any(nd in n for nd in no_decay)
-                        and not any(nd in n for nd in group_all)
-                    ],
-                    'weight_decay_rate': self.params.weight_decay,
-                },
-                {
-                    'params': [
-                        p
-                        for n, p in self.student.named_parameters()
-                        if any(nd in n for nd in no_decay)
-                        and not any(nd in n for nd in group_all)
-                    ],
-                    'weight_decay_rate': 0.0,
-                },
-            ]
-            optimizer_grouped_parameters = (
-                no_decay_optimizer_parameters
-                + decay_optimizer_parameters
-                + group_all_parameters
-            )
-        else:
-            optimizer_grouped_parameters = [
-                {
-                    'params': [
-                        p
-                        for n, p in self.student.named_parameters()
-                        if not any(nd in n for nd in no_decay)
-                    ],
-                    'weight_decay_rate': self.params.weight_decay,
-                },
-                {
-                    'params': [
-                        p
-                        for n, p in self.student.named_parameters()
-                        if any(nd in n for nd in no_decay)
-                    ],
-                    'weight_decay_rate': 0.0,
-                },
-            ]
-        return optimizer_grouped_parameters
-
-    def configure_optimizers(self) -> Tuple[List, List]:
-        """
-        Method to define optimizers and learning rate schedulers
-
-        Returns:
-            Tuple[List, List]: a tuple of containing a list of optimizers and
-                               a list of schedulers to use during training
-        """
-        optimizer_parameters = self._get_student_parameters()
-        if self.params.optimizer == "sgd":
-            optimizer = torch.optim.SGD(
-                optimizer_parameters, lr=self.params.learning_rates[0]
-            )
-        elif self.params.optimizer == "adamw":
-            optimizer = AdamW(
-                optimizer_parameters,
-                lr=self.params.learning_rates[0],
-                eps=self.params.adam_eps,
-            )
-        elif self.params.optimizer == "bertadam":
-            optimizer = BertAdam(
-                optimizer_parameters,
-                lr=self.params.learning_rates[0],
-                warmup=self.params.warmup_ratio,
-            )
-        elif self.params.optimizer == "adam":
-            optimizer = torch.optim.Adam(
-                optimizer_parameters, lr=self.params.learning_rates[0]
-            )
-        else:
-            raise ValueError(f"Optimizer '{self.params.optimizer}' not supported.")
-
-        if self.params.lr_scheduler:
-            scheduler = ReduceLROnPlateau(optimizer)
-            lr_scheduler = {
-                "scheduler": scheduler,
-                "name": "NeptuneLogger",
-                "monitor": "loss",
-            }
-            return [optimizer], [lr_scheduler]
-
-        return [optimizer], []
-
     def get_teacher_logits(self, batch: Dict[str, torch.Tensor]) -> Any:
         raise NotImplementedError()
 
     def get_student_logits(self, batch: Dict[str, torch.Tensor]) -> Any:
-        raise NotImplementedError()
-
-    def training_step(self, batch, _) -> torch.Tensor:
-        raise NotImplementedError()
-
-    def test_step(self, batch, _) -> Dict:
-        raise NotImplementedError()
-
-    def validation_step(self, batch, _) -> Dict:
-        raise NotImplementedError()
-
-    def on_train_epoch_end(self) -> None:
-        """"""
-        self.s_scorer.reset()
-
-    def on_validation_epoch_end(self) -> None:
-        raise NotImplementedError()
-
-    def on_test_epoch_end(self) -> None:
-        raise NotImplementedError()
-
-    def loss(
-        self, teacher_logits: torch.Tensor, student_logits: torch.Tensor, *args, **kwargs
-    ) -> DistillationLoss:
         raise NotImplementedError()
 
     def log_eval_report(self, probs: List[np.array]) -> None:
@@ -294,23 +112,7 @@ class BaseDistiller(pl.LightningModule):
             probs (List[np.array]):
                 predicted probabilities
         """
-        table = self.s_valid_scorer.get_table()
-        self.logger.experiment.add_text("eval/report", table)
-
-        # logging losses to neptune
-        logging_loss = {
-            key: torch.stack(val).mean()
-            for key, val in self.s_valid_scorer.losses.items()
-        }
-        self.log_dict({f"eval/loss_{key}": val for key, val in logging_loss.items()})
-
-        # logging other metrics
-        eval_report = self.s_valid_scorer.to_dict()
-        for key, value in eval_report.items():
-            if not isinstance(value, list) and not isinstance(value, np.ndarray):
-                self.log_dict(
-                    {f"eval/loss_{key}": val for key, val in logging_loss.items()}
-                )
+        super().log_eval_report()
 
         # logging probability distributions
         for i in range(len(probs)):
@@ -321,9 +123,9 @@ class BaseDistiller(pl.LightningModule):
             plt.close("all")
 
 
-class Distiller(BaseDistiller):
+class SequenceClassificationDistiller(BaseSequenceClassificationDistiller):
     """
-    Lightning module to distil a given teacher model into a given student one.
+    Lightning module to distil a given teacher model into a given student one for sequence classification task.
 
     Args:
         teacher (Union["pl.LightningModule", "torch.nn.Module"]):
@@ -337,12 +139,12 @@ class Distiller(BaseDistiller):
     """
 
     def __init__(
-        self,
-        teacher: Union["pl.LightningModule", "torch.nn.Module"],
-        student: Union["pl.LightningModule", "torch.nn.Module"],
-        training_config: DictConfig,
-        teacher_checkpoint: str = None,
-        **kwargs,
+            self,
+            teacher: Union["pl.LightningModule", "torch.nn.Module"],
+            student: Union["pl.LightningModule", "torch.nn.Module"],
+            training_config: DictConfig,
+            teacher_checkpoint: str = None,
+            **kwargs,
     ):
         super().__init__(teacher, student, training_config, teacher_checkpoint, **kwargs)
 
@@ -398,13 +200,13 @@ class Distiller(BaseDistiller):
 
     @overrides
     def loss(
-        self,
-        teacher_logits: torch.Tensor,
-        student_logits: torch.Tensor,
-        labels: torch.Tensor = None,
-        ignore_index: int = -100,
-        *args,
-        **kwargs,
+            self,
+            teacher_logits: torch.Tensor,
+            student_logits: torch.Tensor,
+            labels: torch.Tensor = None,
+            ignore_index: int = -100,
+            *args,
+            **kwargs,
     ) -> DistillationLoss:
         """
         Method called for loss computation
@@ -505,9 +307,12 @@ class Distiller(BaseDistiller):
         self.s_test_scorer.reset()
 
 
-class ParallelDistiller(BaseDistiller):
+class SequenceClassificationParallelDistiller(BaseSequenceClassificationDistiller):
     """
-    Lightning module to distil a given teacher model into a given student one.
+    Lightning module to distil a given teacher model into a given student one for sequence classification
+    tasks.
+
+    The main goal of such class is to distil the teacher's multilingual ability to the student.
 
     Args:
         teacher (Union["pl.LightningModule", "torch.nn.Module"]):
@@ -521,12 +326,12 @@ class ParallelDistiller(BaseDistiller):
     """
 
     def __init__(
-        self,
-        teacher: Union["pl.LightningModule", "torch.nn.Module"],
-        student: Union["pl.LightningModule", "torch.nn.Module"],
-        training_config: DictConfig,
-        teacher_checkpoint: str = None,
-        **kwargs,
+            self,
+            teacher: Union["pl.LightningModule", "torch.nn.Module"],
+            student: Union["pl.LightningModule", "torch.nn.Module"],
+            training_config: DictConfig,
+            teacher_checkpoint: str = None,
+            **kwargs,
     ):
         super().__init__(teacher, student, training_config, teacher_checkpoint, **kwargs)
 
@@ -557,7 +362,7 @@ class ParallelDistiller(BaseDistiller):
 
     @overrides
     def get_student_logits(
-        self, batch: Dict[str, torch.Tensor]
+            self, batch: Dict[str, torch.Tensor]
     ) -> Tuple[torch.Tensor, torch.Tensor]:
         """
         Get student's predictions.
@@ -599,12 +404,12 @@ class ParallelDistiller(BaseDistiller):
 
     @overrides
     def loss(
-        self,
-        teacher_logits: torch.Tensor,
-        student_logits: torch.Tensor,
-        student_logits_translation: torch.Tensor,
-        *args,
-        **kwargs,
+            self,
+            teacher_logits: torch.Tensor,
+            student_logits: torch.Tensor,
+            student_logits_translation: torch.Tensor,
+            *args,
+            **kwargs,
     ) -> DistillationLoss:
         """
         Method called for loss computation, it computes the error between the teacher's and
